Enhance task management system: Introduce new database schemas for user-related data and task dependencies, including migration scripts for existing tables. Update package.json with new dependencies for type generation and UI components. Refactor task-related actions to utilize new schemas and improve data handling. Add new components for task dependencies and enhance existing task forms for better user experience.

+++ database/tables/22_level_settings.sql
@@ -45,8 +45,8 @@ level_exp as (
   select 
     1 as level,
     0 as required_exp,
-    50 as base_exp_for_next_tier, -- 1-10レベルの基本経験値増分
-    1 as tier_multiplier -- 現在の階層の経験値倍率
+    50::numeric as base_exp_for_next_tier, -- 1-10レベルの基本経験値増分
+    1::numeric as tier_multiplier -- 現在の階層の経験値倍率
   
   union all
